abstract UserTable
create component UserTable.tsx from Profiles.tsx
